@@ -1,7 +1,7 @@
 {
   "config": {
     "platform": {
-      "php": "7.0.24"
+      "php": "7.2.2"
     }
   },
   "require-dev": {
@@ -24,7 +24,8 @@
     "symfony/dependency-injection": "^3.4",
     "symfony/config": "^3.4",
     "symfony/http-foundation": "^3.4",
-    "twig/twig": "^2.4"
+    "twig/twig": "^2.4",
+    "ddeboer/imap": "^1.3"
   },
   "autoload": {
     "psr-4": {"Foodsharing\\": "src/",
