@@ -16,7 +16,7 @@ define('DB_PASS', 'root');
 define('DB_DB', 'foodsharing');
 define('PREFIX', 'fs_');
 define('ERROR_REPORT', E_ALL);
-define('BASE_URL', $protocol . '//lmr.local/');
+define('BASE_URL', $protocol . '://lmr.local/');
 define('URL_INTERN', $protocol . '://lmr.local/freiwillige/');
 define('DEFAULT_EMAIL', 'noreply@lebensmittelretten.de');
 define('DEFAULT_EMAIL_NAME', 'Foodsharing Freiwillige');
@@ -26,6 +26,8 @@ define('EMAIL_PUBLIC_NAME', 'Foodsharing Freiwillige');
 define('DEFAULT_HOST', 'lebensmittelretten.de');
 define('MAPZEN_API_KEY', 'mapzen-RaXru7A');
 define('GOOGLE_API_KEY', '');
+
+define('MAILBOX_OWN_DOMAINS', array('lebensmittelretten.de'));
 
 define('SMTP_HOST', 'maildev');
 define('SMTP_PORT', 25);
